Refactor query example rest app to exclude consumer bindings
