+++ activiti-cloud-examples/activiti-cloud-query/config/src/main/resources/activiti-cloud-query-config.properties
@@ -24,19 +24,11 @@ springdoc.swagger-ui.oauth2-redirect-url=/query/swagger-ui/oauth2-redirect.html
 # QUERY
 spring.cloud.stream.bindings.producer.destination=${ACT_QUERY_PRODUCER_DEST:engineEvents}
 spring.cloud.stream.bindings.producer.contentType=${ACT_QUERY_PRODUCER_CONTENT_TYPE:application/json}
-spring.cloud.stream.bindings.queryConsumer.destination=${ACT_QUERY_CONSUMER_DEST:engineEvents}
-spring.cloud.stream.bindings.queryConsumer.group=${ACT_QUERY_CONSUMER_GROUP:query}
-spring.cloud.stream.bindings.queryConsumer.contentType=${ACT_QUERY_CONSUMER_CONTENT_TYPE:application/json}
-
-# AUDIT
-spring.cloud.stream.bindings.auditConsumer.destination=${ACT_AUDIT_CONSUMER_DEST:engineEvents}
-spring.cloud.stream.bindings.auditConsumer.group=${ACT_AUDIT_CONSUMER_GROUP:audit}
-spring.cloud.stream.bindings.auditConsumer.contentType=${ACT_AUDIT_CONSUMER_CONTENT_TYPE:application/json}
 
 activiti.cloud.application.name=default-app
 
 activiti.cloud.messaging.function-router.enabled=false
-activiti.cloud.messaging.function-router.group=${spring.application.name}
+activiti.cloud.messaging.function-router.group=consumer
 activiti.cloud.messaging.function-router.consumer.concurrency=${ACT_QUERY_CONSUMER_CONCURRENCY:1}
 activiti.cloud.messaging.function-router.routes.auditConsumer.enabled=true
 activiti.cloud.messaging.function-router.routes.queryConsumer.enabled=true
